+++ src/setup-docker/index.ts
@@ -150,6 +150,8 @@ export default class SetupDockerGenerator extends Generator {
       fabloBuild: getBuildInfo(),
       fabloRestVersion: "0.1.0",
       hyperledgerExplorerVersion: "1.1.8",
+      fabricCouchDbVersion: "0.4.18",
+      couchDbVersion: "3.1",
       fabricCaPostgresVersion: "14",
     };
     this.fs.copyTpl(this.templatePath("fabric-docker/.env"), this.destinationPath("fabric-docker/.env"), settings);